+++ app/build.gradle.kts
@@ -47,7 +47,4 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-
-    implementation ("org.aviran.cookiebar2:cookiebar2:1.1.4")
-
 }
