+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,7 @@
         android:theme="@style/Theme.Hash_test"
         tools:targetApi="31">
         <activity
-            android:name=".ResultActivity"
-            android:exported="false" />
-        <activity
-            android:name=".HashGenerateActivity"
+            android:name=".Hash_Generate_Activity"
             android:exported="false" />
         <activity
             android:name=".SplashScreen"
