+++ app/src/main/res/layout/activity_hash_generate.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/dark_blue"
-    tools:context=".HashGenerateActivity">
+    tools:context=".Hash_Generate_Activity">
 
     <TextView
         android:id="@+id/textView"
@@ -90,6 +90,19 @@
 
     </com.google.android.material.button.MaterialButton>
 
+    <TextView
+        android:id="@+id/Result"
+        android:layout_width="wrap_content"
+        android:textColor="@color/main_green"
+        android:layout_margin="20dp"
+        android:padding="14dp"
+        android:fontFamily="@font/semi_bold"
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toTopOf="@+id/generate"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/planeText">
 
+    </TextView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
